pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labos"
version = "0.1.0"
description = "LabOS — Self-evolving multi-agent LLM framework for biomedical research"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "LabOS Contributors"},
]
keywords = ["biomedical", "multi-agent", "llm", "research", "self-evolving"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "smolagents>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "markdownify>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "jinja2>=3.1.0",
    "gradio>=4.0.0",
    "tqdm>=4.65.0",
    "biopython>=1.81",
    "pymed>=0.8.9",
    "arxiv>=2.0.0",
]

[project.optional-dependencies]
mem0 = ["mem0ai>=0.1.0"]
screening = ["rdkit-pypi>=2022.9.0", "rapidfuzz>=3.0.0"]
all = [
    "mem0ai>=0.1.0",
    "rdkit-pypi>=2022.9.0",
    "rapidfuzz>=3.0.0",
    "googlesearch-python>=1.2.0",
    "scholarly>=1.7.0",
    "torch>=2.0.0",
]

[project.scripts]
labos = "labos.cli:main"

[project.urls]
Homepage = "https://github.com/labos-ai/labos"
Documentation = "https://github.com/labos-ai/labos#readme"
Repository = "https://github.com/labos-ai/labos"
Issues = "https://github.com/labos-ai/labos/issues"

[tool.setuptools.packages.find]
include = ["labos*"]
